Add files via upload
The python script is designed to parse a log file in which it is used to implement and extract relevant data from it. The data is then extracted to an CSV file format. The Log file contains the connection details of the particular log file in which it is used.

We used some libraries like ‘re’, and ‘csv’. It is used for the regular expression and the CSV file handling. The first of the program is starts with the opening of the log file and the to read the log file.
It defines the regular expression pattern to match and extract data in which it is used to implement from each log file. In the code ‘log_pattern’ is generally deals with the capturing of various connection details like date, time, IDs, rules, protocols, source and destination IP, address ports, zones, bytes, and duration.

The empty list called “log_data” it stores the extracted data from the log file. Then script iterates through each value of the data line and it will sort the data according to the
 
“log_pattern” that we have described in the expression by using the library “re”. If the match is found on a line, the extracted data is appended as a tuple in the “log_data”.

After it iterates through the entire log file it opens a CSV file name format called “outputLog.csv” by using the CSV library. Then it creates the CSV writer object “csv_writer” to write dump the data in the CSV file format.

It writes the data from the “log_data” row-by-row format and it will records all the logs that are matched in the given format of the data in which it is used to sort the each entry of the log file and it will become easier for the Cybersecurity analyst that making it in to the CSV format is much more convenient and making it is accessible for the further processing.